Speed up schedule parsing by avoiding getElementsByTagName()
On my laptop, parsing the schedule took around 800ms. Profiling revealed
that most of the time was spent in getElementsByTagName(), which
traverses the entire subtree for matching elements! Replacing calls to
this function with explicit iteration over the immediate children makes
a huge difference. We now only traverse the document once. This cuts the
time spent parsing down to under 400ms.

(On the N900, parsing took about 5 seconds out of the 8 second startup
time, and now it takes around 2.5 seconds. I haven't formally profiled
there. It seems that the Python startup time is over a second, so this
puts a lower bound on how fast we can launch.)