Don't strip newlines outside descriptions
Tasty over-profiling!

In practice there's only ever 0 or 1 child node, but profiling shows
that the difference is negligible, at 3% or so, and technically we can't
assume this unless we normalize() the document, which is incredibly
expensive. But avoiding the extra split/replace/join makes a slightly
bigger difference, and makes my inner pedant less upset.