Updater: run everything in the worker thread
Previously only the HTTP download was run in a thread; everything else
(deleting the temporary file, copying the new file, etc.) was done in
the main thread.

This made the code a bit harder to follow; it also won't fly if we want
to parse the schedule in a thread, too.